# Optimize conversion of MouseGesture from/to string, reduce allocations