--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Command/MouseGestureConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Command/MouseGestureConverter.cs
@@ -1,14 +1,5 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-//
-//
-//
-// Description: MouseGestureConverter - Converts a MouseGesture string 
-//              to the *Type* that the string represents 
-//
-//
 
 using System.ComponentModel;    // for TypeConverter
 using System.Globalization;     // for CultureInfo
@@ -16,7 +7,7 @@ using System.Globalization;     // for CultureInfo
 namespace System.Windows.Input
 {
     /// <summary>
-    /// MouseGesture - Converter class for converting between a string and the Type of a MouseGesture
+    /// Converter class for converting between a <see cref="string"/> and <see cref="MouseGesture"/>.
     /// </summary>
     public class MouseGestureConverter : TypeConverter
     {