@ThomasGoulet73 2025-01-17
> There is a behavioral change related to the fact that type converters of `MouseAction` and `ModifierKeys` are no longer retrieved via `TypeDescriptor.GetConverter` but rather used statically, which is aligned with how `KeyGestureConverter` and most of other converters using other types that have converters are implemented. This provides around 80-90% of the performance benefit, allows use of `ReadOnlySpan<char>` to pass slices to the converter methods (as you cannot box it in `object`), etc.

Would it be possible to keep the existing behavior by first checking that the type converters registered for `MouseAction` and `ModifierKeys` is the default one ? You could then call the normal methods when there's a custom type converter and call the span methods when it's the default type converter. It would be a bit slower due to the fact that we need to check the registered type converters but at least this change wouldn't be potentially breaking.